HTML website about Brisbane, with a bit of CSS.
Apologies for the extended absence! This only took a day to do, the solution is I need to work more frequently. In this assignment I came across to-be-expected formatting problems, such as getting paragraph text to correspond to the right image and discovered the use of Float and Clear CSS properties. I greatly look forward to learning more about CSS layout!